Cleanup the code, fix minor glitches
