+++ DuckDuckGo/TabSwitcherTransition.swift
@@ -19,12 +19,65 @@
 
 import UIKit
 
-class TabSwitcherTransition: NSObject, UIViewControllerTransitioningDelegate {
+class TabSwitcherTransition: NSObject, UIViewControllerAnimatedTransitioning {
     
     struct Constants {
         static let duration = 0.3
     }
+    
+    // Used to hide contents of the 'from' VC when animating.
+    let solidBackground = UIView()
+    // Container for the image, will clip subviews like tab switcher cell does.
+    let imageContainer = UIView()
+    // Image to display as a preview.
+    let imageView = UIImageView()
+    
+    let tabSwitcherViewController: TabSwitcherViewController
+    
+    init(tabSwitcherViewController: TabSwitcherViewController) {
+        self.tabSwitcherViewController = tabSwitcherViewController
+    }
+    
+    func prepareSubviews(using transitionContext: UIViewControllerContextTransitioning) {
+        
+        transitionContext.containerView.addSubview(solidBackground)
 
+        imageContainer.clipsToBounds = true
+        imageContainer.addSubview(imageView)
+        transitionContext.containerView.addSubview(imageContainer)
+    }
+    
+    // MARK: UIViewControllerAnimatedTransitioning
+
+    // Override - Abstract function
+    func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
+        assertionFailure("You must implement this method")
+    }
+    
+    func transitionDuration(using transitionContext: UIViewControllerContextTransitioning?) -> TimeInterval {
+        return TabSwitcherTransition.Constants.duration
+    }
+    
+    // MARK: Common logic
+    
+    func scrollIfOutsideViewport(collectionView: UICollectionView,
+                                 rowIndex: Int,
+                                 attributes: UICollectionViewLayoutAttributes) {
+        // If cell is outside viewport, scroll while animating
+        if attributes.frame.origin.y + attributes.frame.size.height < collectionView.contentOffset.y {
+            collectionView.scrollToItem(at: IndexPath(row: rowIndex, section: 0),
+                                        at: .top,
+                                        animated: true)
+        } else if attributes.frame.origin.y > collectionView.frame.height + collectionView.contentOffset.y {
+            collectionView.scrollToItem(at: IndexPath(row: rowIndex, section: 0),
+                                        at: .bottom,
+                                        animated: true)
+        }
+    }
+}
+
+class TabSwitcherTransitionDelegate: NSObject, UIViewControllerTransitioningDelegate {
+    
     func animationController(forPresented presented: UIViewController,
                              presenting: UIViewController,
                              source: UIViewController) -> UIViewControllerAnimatedTransitioning? {
@@ -32,13 +85,14 @@ class TabSwitcherTransition: NSObject, UIViewControllerTransitioningDelegate {
             let tabSwitcherVC = presented as? TabSwitcherViewController else {
             return nil
         }
+        
         if mainVC.homeController != nil {
             return FromHomeScreenTransition(mainViewController: mainVC,
                                             tabSwitcherViewController: tabSwitcherVC)
         }
         
-        return TabSwitcherTransitioningIn(mainViewController: mainVC,
-        tabSwitcherViewController: tabSwitcherVC)
+        return FromWebViewTransition(mainViewController: mainVC,
+                                     tabSwitcherViewController: tabSwitcherVC)
     }
 
     func animationController(forDismissed dismissed: UIViewController) -> UIViewControllerAnimatedTransitioning? {
@@ -47,6 +101,6 @@ class TabSwitcherTransition: NSObject, UIViewControllerTransitioningDelegate {
         if let tab = tabSwitcherVC.tabsModel.currentTab, tab.link == nil {
             return ToHomeScreenTransition(tabSwitcherViewController: tabSwitcherVC)
         }
-        return TabSwitcherTransitioningOut(tabSwitcherViewController: tabSwitcherVC)
+        return ToWebViewTransition(tabSwitcherViewController: tabSwitcherVC)
     }
 }
