+++ DuckDuckGo/WebViewTransition.swift
@@ -1,0 +1,168 @@
+//
+//  WebViewTransition.swift
+//  DuckDuckGo
+//
+//  Copyright © 2020 DuckDuckGo. All rights reserved.
+//
+//  Licensed under the Apache License, Version 2.0 (the "License");
+//  you may not use this file except in compliance with the License.
+//  You may obtain a copy of the License at
+//
+//  http://www.apache.org/licenses/LICENSE-2.0
+//
+//  Unless required by applicable law or agreed to in writing, software
+//  distributed under the License is distributed on an "AS IS" BASIS,
+//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//  See the License for the specific language governing permissions and
+//  limitations under the License.
+//
+
+import Core
+
+class WebViewTransition: TabSwitcherTransition {
+    
+    fileprivate func tabSwitcherCellFrame(for attributes: UICollectionViewLayoutAttributes) -> CGRect {
+        var targetFrame = self.tabSwitcherViewController.collectionView.convert(attributes.frame,
+                                                                                to: self.tabSwitcherViewController.view)
+        
+        targetFrame = targetFrame.insetBy(dx: TabViewCell.Constants.cellShadowMargin,
+                                          dy: TabViewCell.Constants.cellShadowMargin)
+        return targetFrame
+    }
+    
+    fileprivate func previewFrame(for cellBounds: CGSize, preview: UIImage) -> CGRect {
+        
+        let previewHeight = cellBounds.width * (preview.size.height / preview.size.width)
+        let targetFrame = CGRect(x: 0,
+                                 y: TabViewCell.Constants.cellHeaderHeight,
+                                 width: cellBounds.width,
+                                 height: previewHeight)
+        return targetFrame
+    }
+}
+
+class FromWebViewTransition: WebViewTransition {
+    
+    private let mainViewController: MainViewController
+    
+    init(mainViewController: MainViewController,
+         tabSwitcherViewController: TabSwitcherViewController) {
+        self.mainViewController = mainViewController
+
+        super.init(tabSwitcherViewController: tabSwitcherViewController)
+    }
+    
+    override func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
+        prepareSubviews(using: transitionContext)
+        tabSwitcherViewController.prepareForPresentation()
+        
+        guard let webView = mainViewController.currentTab!.webView,
+            let tab = mainViewController.currentTab?.tabModel,
+        let rowIndex = tabSwitcherViewController.tabsModel.indexOf(tab: tab),
+        let layoutAttr = tabSwitcherViewController.collectionView.layoutAttributesForItem(at: IndexPath(row: rowIndex, section: 0)),
+        let preview = tabSwitcherViewController.previewsSource.preview(for: tab)
+            else {
+                return }
+        
+        let theme = ThemeManager.shared.currentTheme
+        let webViewFrame = webView.convert(webView.bounds, to: nil)
+        
+        solidBackground.backgroundColor = theme.backgroundColor
+        solidBackground.frame = webViewFrame
+        
+        tabSwitcherViewController.view.alpha = 0
+        transitionContext.containerView.insertSubview(tabSwitcherViewController.view, aboveSubview: solidBackground)
+        tabSwitcherViewController.view.frame = transitionContext.finalFrame(for: tabSwitcherViewController)
+        
+        imageContainer.frame = webViewFrame
+        imageView.frame = imageContainer.bounds
+        imageView.image = preview
+        
+        UIView.animateKeyframes(withDuration: TabSwitcherTransition.Constants.duration, delay: 0, options: .calculationModeLinear, animations: {
+            
+            UIView.addKeyframe(withRelativeStartTime: 0, relativeDuration: 1.0) {
+                let containerFrame = self.tabSwitcherCellFrame(for: layoutAttr)
+                self.imageContainer.frame = containerFrame
+                self.imageContainer.layer.cornerRadius = TabViewCell.Constants.cellCornerRadius
+                self.imageView.frame = self.previewFrame(for: containerFrame.size, preview: preview)
+            }
+            
+            UIView.addKeyframe(withRelativeStartTime: 0.3, relativeDuration: 0.7) {
+                self.tabSwitcherViewController.view.alpha = 1
+            }
+        }, completion: { _ in
+            self.solidBackground.removeFromSuperview()
+            self.imageContainer.removeFromSuperview()
+            transitionContext.completeTransition(true)
+        })
+
+    }
+}
+
+class ToWebViewTransition: WebViewTransition {
+
+    override func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
+        prepareSubviews(using: transitionContext)
+        
+        guard let mainViewController = transitionContext.viewController(forKey: .to) as? MainViewController,
+            let webView = mainViewController.currentTab!.webView,
+            let tab = mainViewController.currentTab?.tabModel,
+            let rowIndex = tabSwitcherViewController.tabsModel.indexOf(tab: tab),
+            let layoutAttr = tabSwitcherViewController.collectionView.layoutAttributesForItem(at: IndexPath(row: rowIndex, section: 0))
+            else {
+                return
+        }
+                
+        let theme = ThemeManager.shared.currentTheme
+        let webViewFrame = webView.convert(webView.bounds, to: nil)
+        mainViewController.view.alpha = 1
+        
+        solidBackground.backgroundColor = theme.backgroundColor
+        solidBackground.frame = webView.bounds
+        // Put overlay above webview to hide its content till the end of the transition
+        solidBackground.removeFromSuperview()
+        webView.addSubview(solidBackground)
+        
+        imageContainer.frame = tabSwitcherCellFrame(for: layoutAttr)
+        imageContainer.layer.cornerRadius = TabViewCell.Constants.cellCornerRadius
+        
+        let preview = tabSwitcherViewController.previewsSource.preview(for: tab)
+        if let preview = preview {
+            imageView.frame = previewFrame(for: imageContainer.bounds.size,
+            preview: preview)
+        } else {
+            imageView.frame = CGRect(origin: .zero, size: imageContainer.bounds.size)
+        }
+        imageView.image = preview
+        
+        scrollIfOutsideViewport(collectionView: tabSwitcherViewController.collectionView, rowIndex: rowIndex, attributes: layoutAttr)
+        
+        UIView.animate(withDuration: TabSwitcherTransition.Constants.duration, animations: {
+            self.imageContainer.frame = webViewFrame
+            self.imageContainer.layer.cornerRadius = 0
+            self.imageView.frame = self.destinationImageFrame(for: webViewFrame.size,
+                                                              preview: preview)
+            
+            self.solidBackground.alpha = 1
+            self.tabSwitcherViewController.view.alpha = 0
+        }, completion: { _ in
+            self.solidBackground.removeFromSuperview()
+            self.imageContainer.removeFromSuperview()
+            transitionContext.completeTransition(true)
+        })
+    }
+    
+    private func destinationImageFrame(for containerSize: CGSize,
+                                       preview: UIImage?) -> CGRect {
+        guard let preview = preview else {
+            return CGRect(origin: .zero, size: containerSize)
+        }
+        
+        let targetFrame = CGRect(x: 0,
+                                 y: 0,
+                                 width: containerSize.width,
+                                 height: containerSize.width * (preview.size.height / preview.size.width))
+        return targetFrame
+    }
+
+}
